refactored /record and /isolated/weight
