+++ apps/gym/src/components/isolated/Weight/WeightList/WeightListEntry/WeightListEntry.tsx
@@ -1,11 +1,12 @@
 import styles from './WeightListEntry.module.css'
+import { WeightListEntryProps } from '@/types'
 import MoreVertIcon from '@mui/icons-material/MoreVert'
 import { EntryModal } from '../../EntryModal'
 import { useToggle } from '@/hooks'
 import { useDeleteWeight } from '@/hooks'
 import { AnimatePresence } from 'framer-motion'
 
-export const WeightListEntry = ({ value, id }) => {
+export const WeightListEntry = ({ value, id }: WeightListEntryProps) => {
   const [isVisible, toggle, _, close] = useToggle()
   const { mutate: deleteWeight } = useDeleteWeight()
