+++ apps/gym/src/types/types.ts
@@ -1,5 +1,5 @@
+import { Dayjs } from 'dayjs'
 import { Dispatch, ReactNode, SetStateAction } from 'react'
-import { ISet } from '../app/(user)/record/record'
 
 export type FormEvent = React.FormEvent<HTMLFormElement>
 export type ButtonEvent = React.MouseEvent<HTMLButtonElement>
@@ -70,3 +70,67 @@ export interface AddWeightModalProps {
   closeModal: () => void
 }
 
+export interface ICurrentSplit {
+  name: string
+}
+
+export interface ISet {
+  weight: string
+  reps: string
+  rpe: string
+}
+
+export interface AddExerciseModalProps {
+  closeModal: () => void
+  workout: IWorkout[]
+  setWorkout: Dispatch<SetStateAction<IWorkout[]>>
+}
+
+export interface CreateNewExerciseModalProps {
+  closeModal: () => void
+}
+
+type ThandleChange = (a: TextInputChangeEvent, b: number, c: number, d: string) => void
+
+type TremoveSet = (a: number, b: number) => void
+
+export interface SetProps {
+  key: number
+  value: ISet
+  exerciseNumber: number
+  setNumber: number
+  handleChange: ThandleChange
+  removeSet: TremoveSet
+}
+
+export interface ExerciseModalProps {
+  toggleExerciseModal: () => void
+  ind: number
+  showNote: boolean
+  toggleNote: () => void
+  workout: IWorkout[]
+  setWorkout: Dispatch<SetStateAction<IWorkout[]>>
+  closeExerciseModal: () => void
+}
+
+export interface WorkoutConfirmationModalProps {
+  closeModal: () => void
+  handleSubmit: (e: ButtonEvent) => void
+}
+
+export interface ChartHeaderProps {
+  startingWeight: number
+  endingWeight?: number
+}
+
+export interface RangeSelectorProps {
+  filter: (startDate?: Dayjs, endDate?: Dayjs) => void
+  openDatePickers: () => void
+  closeDatePickers: () => void
+  data: IWeightData[]
+}
+
+export interface WeightListEntryProps {
+  value: IWeightData
+  id: number
+}
